+++ service-worker.js
@@ -1,5 +1,5 @@
 
-const CACHE="tbop-v14.0.4";
+const CACHE="tbop-v15.0.0-rc1";
 const ASSETS=[
   "./","./index.html","./calendar.html","./membership.html",
   "./login.html","./member-login.html","./officer-login.html",